Move recurse_dependencies to waf_conf.py
